LibWeb: Move StyleFunctionRule code into StyleFunctionRule.cpp
